explorer now gets updated; also now separate files
