+++ openclonk-extension/src/Ifaces/IC4GroupProvider.ts
@@ -1,0 +1,5 @@
+
+export interface IC4groupProvider {
+	unpack(pathToFolder: string): Thenable<void>;
+	pack(pathToFolder: string): Thenable<void>;
+}
